requirements/base.txt
#Python
# -----------------------------------------------------------------------------
pytz==2021.1  # https://github.com/stub42/pytz
python-slugify==5.0.2  # https://github.com/un33k/python-slugify
Pillow==8.2.0  # https://github.com/python-pillow/Pillow
rcssmin==1.0.6  # https://github.com/ndparker/rcssmin
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
redis==3.5.3  # https://github.com/andymccurdy/redis-py
hiredis==1.1.0  # https://github.com/redis/hiredis-py
uvicorn[standard]==0.13.4  # https://github.com/encode/uvicorn
PyJWT==1.7.1
six==1.14.0
packaging==20.9
psutil==5.8.0
topicaxis-opengraph==0.5
grabicon==0.1.2
timeago==1.0.15


# Django
# ------------------------------------------------------------------------------
django==3.1.12  # pyup: < 3.2  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.44.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.11.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-compressor==2.4.1  # https://github.com/django-compressor/django-compressor
django-redis==4.12.1  # https://github.com/jazzband/django-redis
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-appconf==1.0.4
django-compressor==2.4.1
django-coverage-plugin==1.8.0

# Django REST Framework
# ------------------------------------------------------------------------------
djangorestframework==3.12.4  # https://github.com/encode/django-rest-framework
django-cors-headers==3.7.0 # https://github.com/adamchainz/django-cors-headers
djangorestframework-simplejwt==4.4.0
django-rest-auth==0.9.5
dj-rest-auth==2.1.7

# Django Settings
# ------------------------------------------------------------------------------
django-storages==1.11.1
django-stubs==1.8.0
django-stubs-ext==0.2.0

# Django Models
# ------------------------------------------------------------------------------
django-model-utils==4.1.1
django-annoying==0.10.6

# Django Views
# ------------------------------------------------------------------------------
django-filter==2.4.0
drf-yasg==1.20.0
django-url-filter==0.3.15

# Oauth
# ------------------------------------------------------------------------------
oauthlib==3.1.1
requests-oauthlib==1.3.0
